refactor(http): extract post handling service
