+++ Sources/ProxyHTTPTransport/HTTPResponseWriter.swift
@@ -11,76 +11,6 @@ package struct HTTPResponseWriter: Sendable {
         self.logger = logger
     }
 
-    package func handleLocalPostHandling(
-        _ handling: LocalPostHandling,
-        on channel: Channel,
-        prefersEventStream: Bool,
-        keepAlive: Bool,
-        requestLog: HTTPHandler.RequestLogContext
-    ) {
-        switch handling {
-        case .initialize(let future, let sessionID, let originalID):
-            future.whenComplete { result in
-                switch result {
-                case .success(let buffer):
-                    var buffer = buffer
-                    guard let data = buffer.readData(length: buffer.readableBytes) else {
-                        sendPlain(
-                            on: channel,
-                            status: .badGateway,
-                            body: "invalid upstream response",
-                            keepAlive: keepAlive,
-                            sessionID: sessionID,
-                            requestLog: requestLog
-                        )
-                        return
-                    }
-                    sendResponseData(
-                        on: channel,
-                        data: data,
-                        prefersEventStream: prefersEventStream,
-                        keepAlive: keepAlive,
-                        sessionID: sessionID,
-                        requestLog: requestLog
-                    )
-                case .failure:
-                    sendMCPError(
-                        on: channel,
-                        id: originalID,
-                        code: -32000,
-                        message: "upstream timeout",
-                        prefersEventStream: prefersEventStream,
-                        keepAlive: keepAlive,
-                        sessionID: sessionID,
-                        requestLog: requestLog
-                    )
-                }
-            }
-
-        case .immediateResponse(let data, let sessionID):
-            sendResponseData(
-                on: channel,
-                data: data,
-                prefersEventStream: prefersEventStream,
-                keepAlive: keepAlive,
-                sessionID: sessionID,
-                requestLog: requestLog
-            )
-
-        case .mcpError(let id, let code, let message, let sessionID):
-            sendMCPError(
-                on: channel,
-                id: id,
-                code: code,
-                message: message,
-                prefersEventStream: prefersEventStream,
-                keepAlive: keepAlive,
-                sessionID: sessionID,
-                requestLog: requestLog
-            )
-        }
-    }
-
     package func sendSingleSSE(
         on channel: Channel,
         data: Data,
